requirements.txt
git+https://github.com/iMaBoLi/Telethon
tgcrypto
requests
redis
aiofiles
PyGithub
aiocron
Pillow
googletrans==3.1.0a0
speedtest-cli
psutil
instagrapi
mega.py
cowpy
selenium
setuptools
ShazamAPI
yt_dlp
bs4
pymongo[srv]
